Sentence To Sentence Generation
The Idea is to first read all of the sentences of a book. Afterwards, the model should be able to write a sentence after it has been fed a previous sentence. 

In other words, if you feed the model "Stars are very bright in the night sky.", the model should write something like, "They shine all across the universe."